requirements.txt
# Web Framework
streamlit==1.28.1
streamlit-option-menu==0.3.6

# Data Processing
pandas
numpy

# Visualization
matplotlib
plotly==5.18.0

# Biology
biopython

# Testing
pytest==7.4.3
pytest-cov==4.1.0

# Performance
psutil==5.9.6
